retirei cli
