+++ pilot.conf
@@ -1,0 +1,3 @@
+[context]
+verbosity = vv
+
