+++ pyproject.toml
@@ -1,56 +1,47 @@
+# Especificações do sistema de build
 [build-system]
-requires = [
-    "setuptools>=69",
-    "wheel",
-    "setuptools_scm[toml]>=8.1.0",
-]
+requires = ["setuptools>=69","wheel","setuptools_scm[toml]>=8.1.0"]
 build-backend = "setuptools.build_meta"
 
+# Informações básicas do projeto
 [project]
 name = "pytrobot"
-dynamic = [
-    "version",
-]
-description = "A short description of the project"
+version = "3.0.9"
+description = "A transactional robot framework for Python"
+license = { text = "MIT" }
+requires-python = ">=3.11"
+readme = { file = "README.md", content-type = "text/markdown" }
+urls = { homepage = "https://github.com/Batchuka/PyTRobot" }
 authors = [
     { name = "Batchuka", email = "matheusrlz@hotmail.com" },
 ]
-requires-python = ">=3.11"
+
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
     "Operating System :: OS Independent",
 ]
+
 dependencies = [
     "tomli_w==1.0.0",
     "setuptools_scm==8.1.0",
+    "pilot>=2.0.0",
     "invoke==2.2.0",
     "boto3==1.26.141",
     "click==8.1.7",
     "setuptools==65.5.0",
+    "cookiecutter==2.1.1"
 ]
 
-[project.readme]
-file = "README.md"
-content-type = "text/markdown"
-
-[project.license]
-text = "MIT"
-
-[project.urls]
-homepage = "https://github.com/Batchuka/PoG-PyTRobot-framework"
-
+# Scripts que podem ser executados a partir do CLI
 [project.scripts]
 trt = "pytrobot.cli.main:program.run"
 
-[tool.setuptools.packages.find]
-exclude = [
-    "tests",
-    "*.tests",
-    "*.tests.*",
-    "tests.*",
-]
+# Configurações do setuptools
+[tool.setuptools]
+packages.find = { exclude = ["tests", "*.tests", "*.tests.*", "tests.*"] }
 
+# Configurações do setuptools_scm para versionamento automático
 [tool.setuptools_scm]
 version_scheme = "post-release"
 local_scheme = "dirty-tag"
